[refactor] #44 각 컴포넌트 내부에서 onchange, onclick 메소드를 넣도록 수정
