+++ FE/src/components/transaction/TransactionAddModal/ContentInput/index.tsx
@@ -1,8 +1,15 @@
-import React from 'react';
+import React, { ChangeEvent } from 'react';
 
 import './index.scss';
 
-const ContentInput = ({ onContentChange }) => {
+const ContentInput = ({ setContent }) => {
+  const onContentChange = (e: ChangeEvent<HTMLInputElement>) => {
+    if (e.target) {
+      const targetElement = e.target as HTMLInputElement;
+      setContent(targetElement.value);
+    }
+  };
+
   return (
     <div className="item memo__input">
       <div className="indicator">메모</div>
